+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed: Spindle temp reporting when running Analog type spindle without rpm reporting
 - Fixed: Fit the gcode viewer to the path's bounding box instead of its max X/Y/Z
 - Fixed: Last character of the current file was sometimes missing in the file viewer
+- Fixed: Disable trackpad being treated as touchscreen on Linux
 
 [2.1.0]
 - Enhancement: Add right-click menu option to clear resume-at-line setting